+++ main.go
@@ -5,6 +5,9 @@ import (
 	"os"
 
 	"github.com/2859044775/Golider/cmd"
+
+	// 注册内置 addon 模块
+	_ "github.com/2859044775/Golider/internal/addon/modules"
 )
 
 func main() {